typewrappers: Drop typecast in libvirt_intWrap()
The libvirt_intWrap() accepts an integer and passes it to
PyLong_FromLong() which accepts a long. It's perfectly okay to
let compiler do that implicitly.

+++ typewrappers.c
@@ -26,7 +26,7 @@ libvirt_buildPyObject(void *cobj,
 PyObject *
 libvirt_intWrap(int val)
 {
-    return PyLong_FromLong((long) val);
+    return PyLong_FromLong(val);
 }
 
 PyObject *
